// src/types/PU_Types/PU_mainTypes.ts
export type ControlPanelItemsTypes = {
    label: string;
    path: string;
    icon?: JSX.Element;
    style?: string;
}

export type PU_paths = {
    label: string;
    path: string;
    icon?: JSX.Element;
}

export type PU_labels = {
    label: string;
    path: string;
    numOfContent: number;
    icon?: JSX.Element;
}

export type PU_footer = {
    label: string;
    path: string;
}